# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
...
        "/analysis/analyze": {
            "post": {
                "tags": [
                    "analysis"
                ],
                "summary": "Analyze Content",
                "description": "Analyze content from URL or text input using LLM",
                "operationId": "analyze_content_analysis_analyze_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/app__analysis__models__AnalyzeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnalysisResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ]
            }
        },
        "/analysis/list": {
            "get": {
                "tags": [
                    "analysis"
                ],
                "summary": "List Analyses",
                "description": "List all saved analysis results",
                "operationId": "list_analyses_analysis_list_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnalysisListResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ]
            }
        },
        "/analysis/file/{analysis_id}": {
            "get": {
                "tags": [
                    "analysis"
                ],
                "summary": "Get Analysis File",
                "description": "Get analysis file content by ID",
                "operationId": "get_analysis_file_analysis_file__analysis_id__get",
                "parameters": [
                    {
                        "name": "analysis_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Analysis Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "analysis"
                ],
                "summary": "Delete Analysis File",
                "description": "Delete analysis file by ID",
                "operationId": "delete_analysis_file_analysis_file__analysis_id__delete",
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ],
                "parameters": [
                    {
                        "name": "analysis_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Analysis Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/analysis/types": {
            "get": {
                "tags": [
                    "analysis"
                ],
                "summary": "Get Analysis Types",
                "description": "Get available analysis types",
                "operationId": "get_analysis_types_analysis_types_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AnalysisListItem": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "analysis_type": {
                        "$ref": "#/components/schemas/AnalysisType"
                    },
                    "content_type": {
                        "$ref": "#/components/schemas/ContentType"
                    },
                    "created": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created"
                    },
                    "source_url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Source Url"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "title",
                    "analysis_type",
                    "content_type",
                    "created"
                ],
                "title": "AnalysisListItem",
                "description": "Summary item for analysis list responses"
            },
            "AnalysisListResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    },
                    "data": {
                        "items": {
                            "$ref": "#/components/schemas/AnalysisListItem"
                        },
                        "type": "array",
                        "title": "Data"
                    }
                },
                "type": "object",
                "required": [
                    "status",
                    "message",
                    "data"
                ],
                "title": "AnalysisListResponse",
                "description": "Response model for analysis list endpoint"
            },
            "AnalysisResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "title": "Status",
                        "description": "Response status"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "description": "Response message"
                    },
                    "data": {
                        "anyOf": [
                            {
                                "additionalProperties": true,
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Data",
                        "description": "Analysis result data"
                    }
                },
                "type": "object",
                "required": [
                    "status",
                    "message"
                ],
                "title": "AnalysisResponse",
                "description": "Standard response model for analysis operations"
            },
            "AnalysisType": {
                "type": "string",
                "enum": [
                    "summary",
                    "extraction",
                    "classification",
                    "sentiment",
                    "keywords",
                    "custom"
                ],
                "title": "AnalysisType",
                "description": "Types of analysis that can be performed"
            },
            "ContentType": {
                "type": "string",
                "enum": [
                    "webpage",
                    "youtube",
                    "text",
                    "document"
                ],
                "title": "ContentType",
                "description": "Type of content being analyzed"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
            "app__analysis__models__AnalyzeRequest": {
                "properties": {
                    "url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Url",
                        "description": "URL to analyze"
                    },
                    "text": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Text",
                        "description": "Text content to analyze"
                    },
                    "analysis_type": {
                        "$ref": "#/components/schemas/AnalysisType",
                        "description": "Type of analysis to perform",
                        "default": "summary"
                    },
                    "custom_prompt": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Custom Prompt",
                        "description": "Custom prompt for analysis (when analysis_type is CUSTOM)"
                    },
                    "model_name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Model Name",
                        "description": "LLM model to use for analysis",
                        "default": "gpt-4o"
                    },
                    "save_result": {
                        "type": "boolean",
                        "title": "Save Result",
                        "description": "Whether to save the analysis result",
                        "default": true
                    }
                },
                "type": "object",
                "title": "AnalyzeRequest",
                "description": "Request model for analysis endpoint",
                "example": {
                    "analysis_type": "summary",
                    "model_name": "gpt-4o",
                    "save_result": true,
                    "url": "https://example.com/article"
                }
            },
...
        },
        "securitySchemes": {
            "APIKeyHeader": {
                "type": "apiKey",
                "in": "header",
                "name": "X-API-Key"
            }
        }
    }
}